migration_tool/requirements.txt
streamlit>=1.32
oracledb>=2.0.0
snowflake-connector-python>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
